# chore: change stream logic