--- a/src/query/service/src/pipelines/builders/builder_delete.rs
+++ b/src/query/service/src/pipelines/builders/builder_delete.rs
@@ -24,7 +24,6 @@ use common_sql::evaluator::CompoundBlockOperator;
 use common_sql::executor::physical_plans::DeleteSource;
 use common_sql::executor::physical_plans::MutationKind;
 use common_sql::gen_mutation_stream_operator;
-use common_sql::TransformStreamKind;
 use common_storages_fuse::operations::MutationBlockPruningContext;
 use common_storages_fuse::operations::TransformSerializeBlock;
 use common_storages_fuse::FuseLazyPartInfo;
@@ -108,17 +107,19 @@ impl PipelineBuilder {
         )?;
         if table.change_tracking_enabled() {
             let func_ctx = self.ctx.get_function_context()?;
-            let (stream, operators) =
-                gen_mutation_stream_operator(table.schema_with_stream(), true)?;
+            let (stream, operators) = gen_mutation_stream_operator(
+                table.schema_with_stream(),
+                table.get_table_info().ident.seq,
+            )?;
             self.main_pipeline
                 .add_transform(|transform_input_port, transform_output_port| {
                     TransformAddStreamColumns::try_create(
                         transform_input_port,
                         transform_output_port,
-                        TransformStreamKind::Mutation(CompoundBlockOperator {
+                        CompoundBlockOperator {
                             operators: operators.clone(),
                             ctx: func_ctx.clone(),
-                        }),
+                        },
                         stream.clone(),
                     )
                 })?;
@@ -132,6 +133,7 @@ impl PipelineBuilder {
                 output,
                 table,
                 cluster_stats_gen.clone(),
+                false,
             )?;
             proc.into_processor()
         })?;