--- a/src/query/service/src/pipelines/builders/builder_delete.rs
+++ b/src/query/service/src/pipelines/builders/builder_delete.rs
@@ -133,7 +133,7 @@ impl PipelineBuilder {
                 output,
                 table,
                 cluster_stats_gen.clone(),
-                false,
+                MutationKind::Delete,
             )?;
             proc.into_processor()
         })?;