--- a/src/query/service/src/pipelines/builders/builder_merge_into.rs
+++ b/src/query/service/src/pipelines/builders/builder_merge_into.rs
@@ -35,6 +35,7 @@ use common_sql::executor::physical_plans::MergeInto;
 use common_sql::executor::physical_plans::MergeIntoAddRowNumber;
 use common_sql::executor::physical_plans::MergeIntoAppendNotMatched;
 use common_sql::executor::physical_plans::MergeIntoSource;
+use common_sql::executor::physical_plans::MutationKind;
 use common_storages_fuse::operations::common::TransformSerializeSegment;
 use common_storages_fuse::operations::MatchedSplitProcessor;
 use common_storages_fuse::operations::MergeIntoNotMatchedProcessor;
@@ -202,7 +203,7 @@ impl PipelineBuilder {
             OutputPort::create(),
             table,
             cluster_stats_gen,
-            false,
+            MutationKind::MergeInto,
         )?;
 
         let pipe_items = vec![
@@ -286,7 +287,7 @@ impl PipelineBuilder {
             OutputPort::create(),
             table,
             cluster_stats_gen.clone(),
-            false,
+            MutationKind::MergeInto,
         )?
         .get_block_builder();
 
@@ -575,7 +576,7 @@ impl PipelineBuilder {
                 OutputPort::create(),
                 table,
                 cluster_stats_gen.clone(),
-                false,
+                MutationKind::MergeInto,
             )?;
             pipe_items.push(serialize_block_transform.into_pipe_item());
         }