--- a/src/query/service/src/pipelines/builders/builder_update.rs
+++ b/src/query/service/src/pipelines/builders/builder_update.rs
@@ -19,7 +19,6 @@ use common_pipeline_sources::EmptySource;
 use common_sql::evaluator::CompoundBlockOperator;
 use common_sql::executor::physical_plans::UpdateSource;
 use common_sql::gen_mutation_stream_operator;
-use common_sql::TransformStreamKind;
 use common_storages_fuse::operations::TransformSerializeBlock;
 use common_storages_fuse::FuseTable;
 
@@ -50,17 +49,19 @@ impl PipelineBuilder {
 
         if table.change_tracking_enabled() {
             let func_ctx = self.ctx.get_function_context()?;
-            let (stream, operators) =
-                gen_mutation_stream_operator(table.schema_with_stream(), true)?;
+            let (stream, operators) = gen_mutation_stream_operator(
+                table.schema_with_stream(),
+                table.get_table_info().ident.seq,
+            )?;
             self.main_pipeline
                 .add_transform(|transform_input_port, transform_output_port| {
                     TransformAddStreamColumns::try_create(
                         transform_input_port,
                         transform_output_port,
-                        TransformStreamKind::Mutation(CompoundBlockOperator {
+                        CompoundBlockOperator {
                             operators: operators.clone(),
                             ctx: func_ctx.clone(),
-                        }),
+                        },
                         stream.clone(),
                     )
                 })?;
@@ -82,6 +83,7 @@ impl PipelineBuilder {
                 output,
                 table,
                 cluster_stats_gen.clone(),
+                false,
             )?;
             proc.into_processor()
         })