--- a/src/query/sql/src/planner/binder/bind_context.rs
+++ b/src/query/sql/src/planner/binder/bind_context.rs
@@ -558,7 +558,7 @@ impl BindContext {
             column.name(),
             column_index,
             Box::new(column.data_type()),
-            Visibility::Visible,
+            Visibility::InVisible,
         )
         .database_name(Some(table.database().to_string()))
         .table_name(Some(table.name().to_string()))