--- a/src/query/storages/fuse/src/operations/append.rs
+++ b/src/query/storages/fuse/src/operations/append.rs
@@ -80,6 +80,7 @@ impl FuseTable {
                 output,
                 self,
                 cluster_stats_gen.clone(),
+                true,
             )?;
             proc.into_processor()
         })?;