… src/query/service/tests/it/storages/fuse/operations/internal_column.rs
… src/query/sql/src/executor/table_read_plan.rs
… src/query/sql/src/planner/binder/internal_column_factory.rs
… src/query/sql/src/planner/binder/project.rs
… src/query/sql/src/planner/binder/table.rs
… src/query/sql/src/planner/stream_column.rs
… src/query/storages/fuse/src/operations/common/processors/sink_commit.rs
… src/query/storages/fuse/src/operations/common/processors/transform_serialize_block.rs
… src/query/storages/fuse/src/operations/compact.rs
… src/query/storages/fuse/src/operations/mutation/compact/compact_source.rs
… src/query/storages/fuse/src/operations/mutation/mutation_source.rs
… src/query/storages/fuse/src/operations/read/fuse_source.rs
… src/query/storages/fuse/src/operations/read/native_data_source_deserializer.rs
… src/query/storages/fuse/src/operations/read/parquet_data_source_deserializer.rs
… src/query/storages/stream/src/stream_table.rs
… tests/sqllogictests/suites/base/09_fuse_engine/09_0011_change_tracking.test