I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

The stream filter:
```sql
Before:
not(is_not_null(_origin_block_id) and _origin_version<base_table_version)

Now:
not(is_not_null(_origin_version) and 
     (contains(_base_block_ids, _origin_block_id) or _origin_version<base_table_version))
```

```sql
mysql> create table t(a int) cluster by(a) change_tracking=true;
Query OK, 0 rows affected (0.04 sec)

mysql> alter table t set options(change_tracking=false);
Query OK, 0 rows affected (0.03 sec)

mysql> insert into t values(1),(2);
Query OK, 2 rows affected (0.03 sec)

mysql> alter table t set options(change_tracking=true);
Query OK, 0 rows affected (0.03 sec)

mysql> insert into t values(3),(4);
Query OK, 2 rows affected (0.04 sec)

mysql> select _origin_version,_origin_block_id,_origin_block_row_num,a from t order by a;
+-----------------+------------------+-----------------------+------+
| _origin_version | _origin_block_id | _origin_block_row_num | a    |
+-----------------+------------------+-----------------------+------+
|            NULL |             NULL |                  NULL |    1 |
|            NULL |             NULL |                  NULL |    2 |
|            NULL |             NULL |                  NULL |    3 |
|            NULL |             NULL |                  NULL |    4 |
+-----------------+------------------+-----------------------+------+
4 rows in set (0.03 sec)
Read 4 rows, 402.00 B in 0.016 sec., 254.5 rows/sec., 24.98 KiB/sec.

mysql> delete from t where a=3;
Query OK, 1 row affected (0.06 sec)

mysql> select _origin_version,_origin_block_id,_origin_block_row_num,a from t order by a;
+-----------------+-----------------------------------------+-----------------------+------+
| _origin_version | _origin_block_id                        | _origin_block_row_num | a    |
+-----------------+-----------------------------------------+-----------------------+------+
|            NULL |                                    NULL |                  NULL |    1 |
|            NULL |                                    NULL |                  NULL |    2 |
|            9114 | -70265286964003531484154567035708304538 |                     0 |    4 |
+-----------------+-----------------------------------------+-----------------------+------+
3 rows in set (0.03 sec)
Read 3 rows, 241.00 B in 0.009 sec., 346.51 rows/sec., 27.18 KiB/sec.

mysql> optimize table t compact;
Query OK, 3 rows affected (0.09 sec)

mysql> select _origin_version,_origin_block_id,_origin_block_row_num,a from t order by a;
+-----------------+-----------------------------------------+-----------------------+------+
| _origin_version | _origin_block_id                        | _origin_block_row_num | a    |
+-----------------+-----------------------------------------+-----------------------+------+
|            9119 |  99415263896031575503393984678632414870 |                     0 |    1 |
|            9119 |  99415263896031575503393984678632414870 |                     1 |    2 |
|            9114 | -70265286964003531484154567035708304538 |                     0 |    4 |
+-----------------+-----------------------------------------+-----------------------+------+
3 rows in set (0.03 sec)
Read 3 rows, 112.00 B in 0.008 sec., 375.51 rows/sec., 13.69 KiB/sec.

mysql> update t set a=3 where _origin_version=9114;
Query OK, 1 rows affected (0.07 sec)

mysql> select _origin_version,_origin_block_id,_origin_block_row_num,a from t order by a;
+-----------------+-----------------------------------------+-----------------------+------+
| _origin_version | _origin_block_id                        | _origin_block_row_num | a    |
+-----------------+-----------------------------------------+-----------------------+------+
|            9119 |  99415263896031575503393984678632414870 |                     0 |    1 |
|            9119 |  99415263896031575503393984678632414870 |                     1 |    2 |
|            9114 | -70265286964003531484154567035708304538 |                     0 |    3 |
+-----------------+-----------------------------------------+-----------------------+------+
3 rows in set (0.02 sec)
Read 3 rows, 112.00 B in 0.005 sec., 597.05 rows/sec., 21.77 KiB/sec.

mysql> insert into t values(5),(6);
Query OK, 2 rows affected (0.04 sec)

mysql> alter table t recluster;
Query OK, 5 rows affected (0.06 sec)

mysql> select _origin_version,_origin_block_id,_origin_block_row_num,a from t order by a;
+-----------------+-----------------------------------------+-----------------------+------+
| _origin_version | _origin_block_id                        | _origin_block_row_num | a    |
+-----------------+-----------------------------------------+-----------------------+------+
|            9119 |  99415263896031575503393984678632414870 |                     0 |    1 |
|            9119 |  99415263896031575503393984678632414870 |                     1 |    2 |
|            9114 | -70265286964003531484154567035708304538 |                     0 |    3 |
|            9131 |  24672370170860277704458166794042158056 |                     0 |    5 |
|            9131 |  24672370170860277704458166794042158056 |                     1 |    6 |
+-----------------+-----------------------------------------+-----------------------+------+
5 rows in set (0.04 sec)
Read 5 rows, 184.00 B in 0.008 sec., 589.88 rows/sec., 21.20 KiB/sec.
```

- Closes #issue

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/datafuselabs/databend/13808)
<!-- Reviewable:end -->